Enable lazy launch for qemu driver
Current persistent memory is not support hotplug,
so we add the lazy launch functionality to the qemu driver
and perpare for adding dax functionality (via persistent memory)
in future.
(the lazy launch functionality = the vm is not launched until startpod)

All lazy functions are moved to the DriverContext interface,
and the LazyDriverContext is removed.

The interface HypervisorDriver adds AsyncDiskBoot() and AsyncNicBoot(),
and when either one of it returns false, the driver should use
the lazy launch functionality to lunch the vm.

When persistent memory is supported, the qemu driver only brings disks
to qemu argument when boot. The Nics are still async boot.

Signed-off-by: Lai Jiangshan <jiangshanlai@gmail.com>