Use stringer to add string conversons for libhyperstart command codes.
Adds a go generate step to the build process to run stringer (and also download
it if necessary). This is used to generate string mappings for hyperstart op
codes to aid in logging.

To support this change, the hyperstart command code constants are changed to an
explicit type (to be compatible with stringer) with the additional benefit of
improving type-safety surrounding functions which deal with hyperstart (you
must explicitely cast to the HyperstartCode uint32 type)

Logging is adjusted in the relevant places to take advantage of this new
capability, printing the command as well as the code number. This greatly aids
in debugging logs, since the numbers are not deterministic within the code due
to iota generation.

The generated hyperstartcode_stringer.go file is added to the git repo to
allow it to continue to build out of the box.