Fix a race which caused containerd to stop responding.
When a virtual machine error occurs leading to a VM shutdown, there is no
explicit denial of access to the vm.Hub channel used to send commands, but once
the VmContext.loop() function exited, nothing existed handling and cancelling
commands on that channel, nor notifying processes to stop writing to it.

This meant that runv could reliably trigger Docker/supervisord lock-ups when
a problem occurred with the VM startup while container commands were being
waited on (most commonly: `docker run <some image> sleep 10`, but having an
incorrect kernel configured in docker).

This commit fixes the problem by removing the duplication of channel logic
between `Vm` and `VmContext`. `Vm.SendGenericOperation` is made internal, and
calls to a wrapper function on `VmContext.SendVmEvent` to enqueue messages to
the context. This separation enables a check for an active event handler to be
performed before enqueuing a new message - and immediately returns an error if
the message cannot be queued allowing the caller to detect a situation where
the VM has died.

Closes #423.