Merge containerd-nslistener and runv-shim
There's a conflict for current architecture, currently for each
container, we have four processes: runv-shim, runv-containerd,
ns-listener and qemu-system-x86. We show pid of runv-shim as runv
container's pid so that the runv-shim can proxy the exit code, signal
and winsize change, but we also need to show pid of nslistener as runv
container's pid so that it can listen on network change so that the CNI
network plugin can work.

Also the future POD support also need one single shim/ns-listener
process pid to identify the container, two processes makes everything a
chaos.

All these things lead to one result: runv-shim and nslistener must belong
to a single process. That's exactly what this commit is doing.

Enjoy it!

Signed-off-by: Zhang Wei <zhangwei555@huawei.com>