Add explicit delete command to the OCI interface
Currently runv automatically destroys any container that
finish executing. The OCI spec defines that the container
shall enter a stopped status instead; an explicit delete command
is defined for cleaning up any container resources.

Don't delete the container automatically on exit, and add an
explicit delete command to the OCI interface of runv. To get
around the limited containerd interface of runv, use a special
status for containers that are about to be deleted.

Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>