api: grpc: Update network API
This commit updates gRPC API related to network so that we can get
an hardware address related to the interface. This will help to
determine the right interface inside the VM as we might not know
the name created inside the VM.

Also, it renames UpdateInterface into AddInterfaces, and AddRoute
into AddRoutes for more consistency. Indeed, we can add more than
one route and more than one interface if needed.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>