api: grpc: Add commands to stop and remove a container
This commit introduces extra commands related to a container, and
specifically to the tear down of a container.

StopContainer() will be used in order to force a container to be
stopped without having to bother finding back the container process.
It will simply kill every processes associated to this container and
finish with the container process.

RemoveContainer() will remove a container from the VM, by tearing down
everything related to this container, so that a new container with the
same ID could be started again later.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>